+++ Flow/src/utopia/flow/collection/immutable/tree/TreeMutatorLike.scala
@@ -1,7 +1,7 @@
 package utopia.flow.collection.immutable.tree
 
 import utopia.flow.collection.CollectionExtensions._
-import utopia.flow.collection.template.tree.{TreeLike2, TreeNavigator}
+import utopia.flow.collection.template.tree.{TreeLike, TreeNavigator}
 import utopia.flow.util.Mutate
 
 import scala.annotation.tailrec
@@ -16,7 +16,7 @@ import scala.annotation.unchecked.uncheckedVariance
  * @author Mikko Hilpinen
  * @since 07.08.2026, v2.9
  */
-trait TreeMutatorLike[-Nav, N, Node <: CopyableTreeLike[N, Node], +Repr <: TreeLike2[Node]]
+trait TreeMutatorLike[-Nav, N, Node <: CopyableTreeLike[N, Node], +Repr <: TreeLike[Node]]
 	extends TreeNavigator[Nav, Repr] with CopyableTreeLike[N, Node]
 {
 	// ABSTRACT   ------------------------